Introduce BlockingStatsDClient analogue to NonBlockingStatsDClient for usecases where no additional executor is required. Common functionality has been extracted to common abstract base classes while preserving all interfaces to keep backwards compatibility.
