+++ src/main/java/com/timgroup/statsd/BlockingStatsDClient.java
@@ -1,0 +1,142 @@
+package com.timgroup.statsd;
+
+import java.net.InetSocketAddress;
+import java.nio.ByteBuffer;
+import java.util.concurrent.Callable;
+
+/**
+ * A simple StatsD client implementation facilitating metrics recording.
+ *
+ * <p>Upon instantiation, this client will establish a socket connection to a StatsD instance
+ * running on the specified host and port. Metrics are then sent over this connection as they are
+ * received by the client.
+ * </p>
+ *
+ * <p>Three key methods are provided for the submission of data-points for the application under
+ * scrutiny:
+ * <ul>
+ *   <li>{@link #incrementCounter} - adds one to the value of the specified named counter</li>
+ *   <li>{@link #recordGaugeValue} - records the latest fixed value for the specified named gauge</li>
+ *   <li>{@link #recordExecutionTime} - records an execution time in milliseconds for the specified named operation</li>
+ *   <li>{@link #recordHistogramValue} - records a value, to be tracked with average, maximum, and percentiles</li>
+ *   <li>{@link #recordEvent} - records an event</li>
+ *   <li>{@link #recordSetValue} - records a value in a set</li>
+ * </ul>
+ * From the perspective of the application, these methods are blocking, with the resulting
+ * IO operations being carried out in the same thread.
+ * The methods may propagate an exception which may disrupt application execution.
+ *
+ * <p>As part of a clean system shutdown, the {@link #stop()} method should be invoked
+ * on any StatsD clients.</p>
+ *
+ * @author Daniel Würsch
+ *
+ */
+public final class BlockingStatsDClient extends AbstractDatagramStatsDClient {
+
+    /**
+     * Create a new StatsD client communicating with a StatsD instance on the
+     * specified host and port. All messages send via this client will have
+     * their keys prefixed with the specified string. The new client will
+     * attempt to open a connection to the StatsD server immediately upon
+     * instantiation, and may throw an exception if that a connection cannot
+     * be established. All exceptions thrown during subsequent usage are
+     * propagated and failure in metrics will affect normal code execution.
+     *
+     * @param prefix
+     *     the prefix to apply to keys sent via this client
+     * @param hostname
+     *     the host name of the targeted StatsD server
+     * @param port
+     *     the port of the targeted StatsD server
+     * @throws StatsDClientException
+     *     if the client could not be started
+     */
+    public BlockingStatsDClient(final String prefix, final String hostname, final int port) throws StatsDClientException {
+        this(prefix, hostname, port, null, null);
+    }
+
+    /**
+     * Create a new StatsD client communicating with a StatsD instance on the
+     * specified host and port. All messages send via this client will have
+     * their keys prefixed with the specified string. The new client will
+     * attempt to open a connection to the StatsD server immediately upon
+     * instantiation, and may throw an exception if that a connection cannot
+     * be established. All exceptions thrown during subsequent usage are
+     * propagated and failure in metrics will affect normal code execution.
+     *
+     * @param prefix
+     *     the prefix to apply to keys sent via this client
+     * @param hostname
+     *     the host name of the targeted StatsD server
+     * @param port
+     *     the port of the targeted StatsD server
+     * @param constantTags
+     *     tags to be added to all content sent
+     * @throws StatsDClientException
+     *     if the client could not be started
+     */
+    public BlockingStatsDClient(final String prefix, final String hostname, final int port, final String... constantTags) throws StatsDClientException {
+        this(prefix, hostname, port, constantTags, null);
+    }
+
+    /**
+     * Create a new StatsD client communicating with a StatsD instance on the
+     * specified host and port. All messages send via this client will have
+     * their keys prefixed with the specified string. The new client will
+     * attempt to open a connection to the StatsD server immediately upon
+     * instantiation, and may throw an exception if that a connection cannot
+     * be established. All exceptions thrown during subsequent usage are
+     * propagated and failure in metrics will affect normal code execution.
+     *
+     * @param prefix
+     *     the prefix to apply to keys sent via this client
+     * @param hostname
+     *     the host name of the targeted StatsD server
+     * @param port
+     *     the port of the targeted StatsD server
+     * @param constantTags
+     *     tags to be added to all content sent
+     * @param errorHandler
+     *     handler to use when an exception occurs during usage, may be null to indicate noop
+     * @throws StatsDClientException
+     *     if the client could not be started
+     */
+    public BlockingStatsDClient(final String prefix, final String hostname, final int port,
+                                final String[] constantTags, final StatsDClientErrorHandler errorHandler) throws StatsDClientException {
+        this(prefix, constantTags, errorHandler, staticStatsDAddressResolution(hostname, port));
+    }
+
+    /**
+     * Create a new StatsD client communicating with a StatsD instance on the
+     * specified host and port. All messages send via this client will have
+     * their keys prefixed with the specified string. The new client will
+     * attempt to open a connection to the StatsD server immediately upon
+     * instantiation, and may throw an exception if that a connection cannot
+     * be established. All exceptions thrown during subsequent usage are
+     * propagated and failure in metrics will affect normal code execution.
+     *
+     * @param prefix
+     *     the prefix to apply to keys sent via this client
+     * @param constantTags
+     *     tags to be added to all content sent
+     * @param errorHandler
+     *     handler to use when an exception occurs during usage, may be null to indicate noop
+     * @param addressLookup
+     *     yields the IP address and socket of the StatsD server
+     * @throws StatsDClientException
+     *     if the client could not be started
+     */
+    public BlockingStatsDClient(final String prefix, String[] constantTags, final StatsDClientErrorHandler errorHandler,
+                                final Callable<InetSocketAddress> addressLookup) throws StatsDClientException {
+        super(prefix, constantTags, errorHandler, addressLookup);
+    }
+
+    void send(final byte[] message) {
+        try {
+            sendBlocking(ByteBuffer.wrap(message));
+        } catch (Exception e) {
+            throw new StatsDClientException("Unable to send", e);
+        }
+    }
+}
